Does simple financial calculation.
This program checks a monthly payment, a loan period, or the total cost of something given the interest rate. If a person is looking for the amount of a monthly payment, then the person must enter a loan period, total cost, and interest rate. If one variable from the first 3 variables is missing this program can be used to find the value of that missing variable.