Game with a smarter AI not finished!
Wanted to make a tic tac toe game with a smarter AI. Instead of using if else statements I wanted to list the winning the combinations, and based on the positions, I wanted the computer to make mathematical calculations to find the best spot to place the next move. Because I was pressed on time and the assignment was due I ended up just using a bunch of if else statements. Hence to say the AI isn't the smartest or the greatest competition.